Restructure project
All classes that are shared between server and client are now ONLY
located in common package
